+++ include/optimizer.h
@@ -22,19 +22,19 @@ struct Optimizer {
 };
 
 typedef struct {
-    double learning_rate;
+    float learning_rate;
 } SGDConfig;
 
 typedef struct {
-    double learning_rate;
-    double beta;
+    float learning_rate;
+    float beta;
     int n_layers;
     Matrix** weight_momentum;
     Matrix** bias_momentum;
 } MomentumConfig;
 
 typedef struct {
-    double learning_rate;
+    float learning_rate;
     int n_layers;
     Matrix** weight_s;
     Matrix** bias_s;
@@ -43,9 +43,9 @@ typedef struct {
 } AdaGradConfig;
 
 typedef struct {
-    double learning_rate;
-    double beta_m;
-    double beta_s;
+    float learning_rate;
+    float beta_m;
+    float beta_s;
     int n_layers;
     int ctr;
 
@@ -62,22 +62,22 @@ typedef struct {
     Matrix** intermediate_b;
 } AdamConfig;
 
-Optimizer* optimizer_sgd_new(double learning_rate);
+Optimizer* optimizer_sgd_new(float learning_rate);
 void optimizer_sgd_free(Optimizer* optimizer);
 void update_weights_sgd(Matrix* weights, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 void update_bias_sgd(Matrix* bias, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 
-Optimizer* optimizer_momentum_new(double learning_rate, double beta, bool nesterov);
+Optimizer* optimizer_momentum_new(float learning_rate, float beta, bool nesterov);
 void optimizer_momentum_free(Optimizer* optimizer);
 void update_weights_momentum(Matrix* weights, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 void update_bias_momentum(Matrix* bias, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 
-Optimizer* optimizer_adagrad_new(double learning_rate);
+Optimizer* optimizer_adagrad_new(float learning_rate);
 void optimizer_adagrad_free(Optimizer* optimizer);
 void update_weights_adagrad(Matrix* weights, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 void update_bias_adagrad(Matrix* bias, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 
-Optimizer* optimizer_adam_new(double learning_rate, double beta_m, double beta_s);
+Optimizer* optimizer_adam_new(float learning_rate, float beta_m, float beta_s);
 void optimizer_adam_free(Optimizer* optimizer);
 void update_weights_adam(Matrix* weights, Matrix* gradient, Optimizer* optimizer, int layer_idx);
 void update_bias_adam(Matrix* bias, Matrix* gradient, Optimizer* optimizer, int layer_idx);
