+++ src/score.c
@@ -19,8 +19,8 @@ void score_free(Score* s) {
 void score_batch(Score* self, Matrix* y_pred, Matrix* y_true) {
     matrix_argmax_into(y_pred, self->y_pred_argmax);
     matrix_argmax_into(y_true, self->y_true_argmax);
-    double n_samples = self->y_pred_argmax->n_cols;
-    double correct_preds = 0.0;
+    float n_samples = self->y_pred_argmax->n_cols;
+    float correct_preds = 0.0;
     for (int i=0; i<n_samples; i++) {
         if (matrix_get(self->y_pred_argmax,0,i) == matrix_get(self->y_true_argmax,0,i)) {
             correct_preds++;
@@ -33,11 +33,11 @@ void score_batch(Score* self, Matrix* y_pred, Matrix* y_true) {
 void update_confusion_matrix(Score* self, Matrix* y_pred, Matrix* y_true, Matrix* c) {
     matrix_argmax_into(y_pred, self->y_pred_argmax);
     matrix_argmax_into(y_true, self->y_true_argmax);
-    double n_samples = self->y_pred_argmax->n_cols;
+    float n_samples = self->y_pred_argmax->n_cols;
     for (int i=0; i<n_samples; i++) {
         int true_class = matrix_get(self->y_true_argmax, 0, i);
         int pred_class = matrix_get(self->y_pred_argmax, 0, i);
-        double count = matrix_get(c, true_class, pred_class);
+        float count = matrix_get(c, true_class, pred_class);
         matrix_assign(c, true_class, pred_class, count+1);
     }
 }
